+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated dependency cfg_if to v1.0
 
+### Fixed
+- Fixed the start position of tokens at the beginnig of a line to not be at the end of the previous line.
+
 ## [0.9.0] - 2020-12-21
 ### Added
 - Added support for retrieving the `Punctuated` sequence of fields in a `TableConstructor`